This python program check two documents weather they are similar or not using multidimensional vectors, sets,hash tables,etc.
The task is to make it as simple as possible. Also the visuaalization of data at the end of the inspection is added.
Other features include the positioning coeficients which  checks the position of words. This feature can be easily fooled and the main task is to enhance this feature so that missing 2-3 words would not cause the positioning coefficient to abnormally lower down.